levo-schema-api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.20
pydantic==2.5.1
python-multipart==0.0.6
pyyaml==6.0
alembic==1.11.1
pytest==7.4.0
requests==2.31.0
httpx==0.24.1
